+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Replace tilde (`~`), environment variables, `${userHome}`, `${workspaceFolder}` in options
 - Replace tidle (`~`) with home directory in `\include`-like commands
+- Add "Go To Definition" support for user-defined commands with `\def` and `\let` ([#1081](https://github.com/latex-lsp/texlab/issues/1081))
 
 ## [5.14.1] - 2024-03-27